Remove linter disable commands and edit README.md to include license.
