+++ MIT.md
@@ -1,0 +1,8 @@
+## Copyright 2022, Amanuel Worku
+
+
+Permission is hereby granted, free of charge, to any person obtaining a copy of this Webpage and associated documentation files, to deal in the Webpage without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Webpage, and to permit persons to whom the Webpage is furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Webpage.
+
+THE WEBPAGE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE WEBPAGE OR THE USE OR OTHER DEALINGS IN THE WEBPAGE.
