+++ README.md
@@ -26,6 +26,7 @@
   - [🔭 Future Features ](#-future-features-)
   - [🤝 Contributing ](#-contributing-)
   - [⭐️ Show your support ](#️-show-your-support-)
+  - [📝 License](#license)
 
 <!-- PROJECT DESCRIPTION -->
 
@@ -130,6 +131,13 @@ Feel free to check the [issues page](https://github.com/AmanWorku/To-Do-List/iss
 
 If you like this project please follow my github account and give a star to the projects you like. It helps a lot. Much Love.
 
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+<!-- LICENSE -->
+
+## 📝 License <a name="license"></a>
+
+This project is [MIT](./MIT.md) licensed.
 <p align="right">(<a href="#readme-top">back to top</a>)</p>
